[refactor]: BusSimulator 위치 변경
softeer5th#58
