+++ backend/src/main/java/com/ddbb/dingdong/simulator/BusSimulatorFactory.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.ddbb.dingdong.simulator.subscription.source.BusSimulator;
 import org.springframework.data.geo.Point;
 import org.springframework.stereotype.Service;
